feat: floating window
